LayoutTests/media/video-currentTime-set-expected.txt

EVENT(load)
EVENT(timeupdate)
EVENT(ended)
END OF TEST
